Moved request/response records to the corresponding auth controller package
