+++ src/main/java/com/rbaun/banking/controller/v1/auth/AuthController.java
@@ -1,7 +1,7 @@
 package com.rbaun.banking.controller.v1.auth;
 
-import com.rbaun.banking.auth.AuthRequest;
-import com.rbaun.banking.auth.JwtResponse;
+import com.rbaun.banking.controller.v1.auth.request.AuthRequest;
+import com.rbaun.banking.controller.v1.auth.response.JwtResponse;
 import com.rbaun.banking.auth.JwtUtil;
 import com.rbaun.banking.auth.UserRepository;
 import com.rbaun.banking.model.user.UserInfo;
