+++ src/main/java/com/rbaun/banking/controller/v1/auth/request/AuthRequest.java
@@ -1,4 +1,4 @@
-package com.rbaun.banking.auth;
+package com.rbaun.banking.controller.v1.auth.request;
 
 import jakarta.validation.constraints.NotBlank;
