+++ src/main/java/com/rbaun/banking/controller/v1/auth/response/JwtResponse.java
@@ -1,4 +1,4 @@
-package com.rbaun.banking.auth;
+package com.rbaun.banking.controller.v1.auth.response;
 
 public record JwtResponse(
         String token
